Improve user preference loading
